Renamed paths.js to index.js and moved it to the service folder
