+++ services/data.service.js
@@ -4,6 +4,6 @@ export async function loadData(path) {
     const data = await response.json();
     return data;
   } catch (err) {
-    return console.log("Error while fetching the data: " + err.message);
+    console.log("Error while fetching the data: " + err.message);
   }
 }
